+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Prop `hideUnavailableVariations` that enables hiding unavailable variations from DOM
+
 ## [3.178.4] - 2025-09-30
 
 ### Fixed
@@ -48,7 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Youtube shorts as a valid format
-  
+
 ## [3.176.3] - 2025-02-28
 
 ### Added